requirements.txt
streamlit
python-dotenv

langchain
langchain-core
langchain-community
langchain-chroma
langchain-openai
langchain-huggingface

chromadb
sentence-transformers